Modify .saveregs and .restoreregs to be able to save/restore vector registers.  These now assume/maintain 64 byte stack alignment. Use these to save state while recursively rasterizing.

+++ tools/assembler/code_output.h
@@ -161,7 +161,7 @@ void align(int alignment);
 void reserve(int amt);
 void emitLabelAddress(const struct Symbol *sym, int lineno);
 int adjustFixups(void);
-void saveRegs(unsigned int bitmask, int lineno);
-void restoreRegs(unsigned int bitmask, int lineno);
+void saveRegs(unsigned long long int bitmask, int lineno);
+void restoreRegs(unsigned long long int bitmask, int lineno);
 
 #endif
